requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.4
python-dotenv==1.0.1
numpy==1.26.4
pydantic==2.10.6
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.0
